requirements.txt
aio-pika
asyncpg
bcrypt~=4.0.1
colorama
httpx
jinja2
pip-chill
python-rapidjson
sanic~=22.9.1
sanic-ext
sanic-httpauth
secure~=0.3.0
tortoise-orm
eth-account
inflect
phonenumbers~=8.13.0
python-dateutil~=2.8.2
PyJWT~=2.6.0